fix(lms): reconstruct the corrupted contracts/lms crate and wire it into the workspace
contracts/lms was excluded from the workspace `members` list and its source
files (course.rs, storage.rs, models.rs, test.rs) contained unresolved merge
debris: multiple full, conflicting copies of the same structs/enums/functions
concatenated together, including literal `=======` diff3 markers. It could
not compile in any configuration.

Reconciled into a single coherent design:
- course.rs owns one canonical `Course` struct (id, instructor, title,
  description, category, difficulty, thumbnail, published, archived,
  created_at, updated_at) and one `CourseError` enum, unifying three
  previously-incompatible versions. All course storage now consistently
  uses persistent storage (previously mixed instance/persistent, which
  would have made lookups fail silently). Courses now start unpublished
  (draft) and require an explicit `publish_course` call, matching what
  the surviving tests actually expected.
- models.rs keeps Module/Lesson/Quiz (deduplicating a copy-pasted second
  `Quiz` definition) and drops its own redundant, incompatible `Course`
  definition in favor of course.rs's.
- storage.rs keeps one `StorageKey` enum (adding `NextCourseId`, needed
  by create_course's id counter) and drops an entirely separate, unused
  `DataKey`/`StudentProfile` subsystem that nothing referenced.
- contract.rs now actually exposes the crate's functions as callable
  `LMSContract` entry points (delegating to course.rs/admin.rs, matching
  the established pattern in contracts/savings), instead of being a
  stub with only `initialize() -> bool { true }`. `initialize` now
  wires through to admin.rs's already-fully-implemented, already-tested
  `initialize_admin`.
- event.rs gains `emit_course_published`/`emit_course_archived` for
  parity with the create/enroll events it already had, and course.rs's
  functions now emit through `LMSEvents` consistently instead of ad hoc
  `env.events().publish(...)` calls.
- The tangled top-level test.rs (which concatenated what were clearly
  meant to be separate per-module test files, several missing their
  enclosing `mod` block or `use super::*`) is removed; its scenarios are
  rewritten as `#[cfg(test)] mod tests` blocks inside the module they
  actually test, matching the convention already used by errors.rs.
  Tests now register a real `LMSContract` and run under
  `env.as_contract(...)`, required by this SDK version for any
  storage/event/auth access; sequential `require_auth()` calls for the
  same address are split into separate `as_contract` frames where
  needed to avoid `Auth::ExistingValue` under `mock_all_auths()`.
- Added contracts/lms to the workspace `members` list.

Note for anyone testing this crate in isolation (`cargo test -p lms` /
`cargo build -p lms`): a fresh, workspace-narrow dependency resolution
can pick `ed25519-dalek 3.0.0`, which breaks `soroban-env-host`'s test
utilities (a `CryptoRng` bound mismatch). The root package's own pin
(`ed25519-dalek = "2.2"`) only takes effect once it's part of the
resolved dependency closure, which a narrowly-scoped build may not
include. If hit, run `cargo update -p ed25519-dalek:3.0.0 --precise
2.2.0`. Cargo.lock is gitignored in this repo, so this isn't fixable by
committing a lockfile; a full `cargo test --workspace` is unaffected.

Verified: `cargo test -p lms` (25/25 passing), `cargo build -p lms
--release` and `--release --target wasm32v1-none` both succeed.
`cargo check --workspace --keep-going` shows the exact same 5
pre-existing, unrelated broken crates (allowances, batch-wallet-
creation, budget, multi-currency-wallet, savings-goals) on this branch
as on main without this change — this fix introduces no new breakage
elsewhere in the workspace.
