+++ Cargo.toml
@@ -44,6 +44,7 @@ members = [
     "contracts/penalty",
     "contracts/transaction-analytics",
     "contracts/treasury",
+    "contracts/lms",
 ]
 
 [package]
