+++ contracts/lms/src/admin.rs
@@ -81,3 +81,86 @@ pub fn require_instructor_or_admin(env: &Env, caller: &Address) -> Result<(), Ad
         Err(AdminError::Unauthorized)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use soroban_sdk::testutils::Address as _;
+
+    #[test]
+    fn test_initialize_and_get_admin() {
+        let env = Env::default();
+        env.mock_all_auths();
+        let contract_id = env.register(crate::LMSContract, ());
+
+        let admin = Address::generate(&env);
+
+        env.as_contract(&contract_id, || {
+            assert!(initialize_admin(env.clone(), admin.clone()).is_ok());
+
+            assert_eq!(get_admin(&env).unwrap(), admin);
+            assert_eq!(get_role(&env, &admin), Role::Admin);
+
+            // Second initialization must fail
+            assert_eq!(
+                initialize_admin(env.clone(), admin.clone()),
+                Err(AdminError::AlreadyInitialized)
+            );
+        });
+    }
+
+    #[test]
+    fn test_authorized_role_assignment() {
+        let env = Env::default();
+        env.mock_all_auths();
+        let contract_id = env.register(crate::LMSContract, ());
+
+        let admin = Address::generate(&env);
+        let instructor = Address::generate(&env);
+
+        // Split into separate frames: a second `require_auth()` for the same
+        // `admin` address within one frame trips `Auth::ExistingValue` under
+        // `mock_all_auths()`.
+        env.as_contract(&contract_id, || {
+            initialize_admin(env.clone(), admin.clone()).unwrap();
+        });
+
+        env.as_contract(&contract_id, || {
+            // Admin assigns Instructor role
+            let result = set_role(env.clone(), admin, instructor.clone(), Role::Instructor);
+            assert!(result.is_ok());
+            assert_eq!(get_role(&env, &instructor), Role::Instructor);
+
+            // Instructor passes instructor/admin authorization check
+            assert!(require_instructor_or_admin(&env, &instructor).is_ok());
+        });
+    }
+
+    #[test]
+    fn test_unauthorized_action_rejected() {
+        let env = Env::default();
+        env.mock_all_auths();
+        let contract_id = env.register(crate::LMSContract, ());
+
+        let admin = Address::generate(&env);
+        let student = Address::generate(&env);
+        let target = Address::generate(&env);
+
+        env.as_contract(&contract_id, || {
+            initialize_admin(env.clone(), admin).unwrap();
+
+            // Default role is Student
+            assert_eq!(get_role(&env, &student), Role::Student);
+
+            // Student attempting to set role must fail
+            let result = set_role(env.clone(), student.clone(), target, Role::Instructor);
+            assert_eq!(result, Err(AdminError::Unauthorized));
+
+            // Student attempting privileged action must fail
+            assert_eq!(
+                require_instructor_or_admin(&env, &student),
+                Err(AdminError::Unauthorized)
+            );
+        });
+    }
+}
