+++ contracts/lms/src/lib.rs
@@ -1,10 +1,11 @@
 #![no_std]
 
+pub mod admin;
 mod contract;
-
-#[cfg(test)]
-mod test;
+pub mod course;
+pub mod errors;
+pub mod event;
+pub mod models;
+pub mod storage;
 
 pub use contract::*;
-
-pub mod errors;
